requirements.txt
python-vlc
pyinstaller